+++ docs/developer.md
@@ -19,7 +19,7 @@ The `LazyBuffer` graph specifies the compute in terms of low level tinygrad ops.
 
 The [scheduler](https://github.com/tinygrad/tinygrad/tree/master/tinygrad/engine/schedule.py) converts the graph of LazyBuffers into a list of `ScheduleItem`. One `ScheduleItem` is one kernel on the GPU, and the scheduler is responsible for breaking the large compute graph into subgraphs that can fit in a kernel. `ast` specifies what compute to run, and `bufs` specifies what buffers to run it on.
 
-::: tinygrad.ops.ScheduleItem
+::: tinygrad.engine.schedule.ScheduleItem
 
 ## Lowering
