+++ openpilot/compile2.py
@@ -19,8 +19,8 @@ from tinygrad.device import Buffer
 from tinygrad.helpers import partition, Context, fetch, getenv, DEBUG
 from tinygrad.engine.realize import run_schedule, lower_schedule, ExecItem, CompiledRunner
 from tinygrad.engine.memory import memory_planner
-from tinygrad.engine.schedule import create_schedule
-from tinygrad.ops import LoadOps, ScheduleItem
+from tinygrad.engine.schedule import ScheduleItem, create_schedule
+from tinygrad.ops import LoadOps
 Device.DEFAULT = "GPU"
 
 def get_schedule(onnx_data) -> Tuple[List[ScheduleItem], List[ScheduleItem]]:
