+++ test/external/fuzz_schedule.py
@@ -5,8 +5,8 @@ from tinygrad.device import Buffer
 from tinygrad.engine.realize import CustomOp, capturing, lower_schedule_item
 from tinygrad.helpers import DEBUG, MULTIOUTPUT, colored, getenv
 from tinygrad.lazy import LazyBuffer
-from tinygrad.engine.schedule import _graph_schedule, _LBScheduleItem
-from tinygrad.ops import LoadOps, ScheduleItem
+from tinygrad.engine.schedule import _graph_schedule, _LBScheduleItem, ScheduleItem
+from tinygrad.ops import LoadOps
 from tinygrad.tensor import Tensor
 
 ctx_vars = { MULTIOUTPUT: (0, 1) }
